# abrg/openapi/openapi.yml
openapi: 3.1.0
info:
  title: ジオコーダーAPI
  version: 3.0.22
  description: |
    ジオコーダーAPIは、アドレス・ベース・レジストリを利用し、住所の正規化およびジオコーディングを行うことが可能です。
servers:
  - url: http://localhost:3000/
paths:
  "/normalize":
    get:
      summary: 住所正規化
      description: 住所を正規化します（表記揺れの統一、DB照合なし）。
      operationId: normalizeAddress
      parameters:
        - $ref: "#/components/parameters/address"
      responses:
        "200":
          description: 正規化結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  input:
                    type: string
                    description: 入力住所
                    example: "東京都千代田区紀尾井町１番３号"
                  output:
                    type: string
                    description: 正規化後の住所
                    example: "東京都千代田区紀尾井町1-3"
                  type:
                    type: string
                    description: |
                      テキストパターンから推測した住所タイプ（DB照合なし）

                      value | description | pattern
                      --- | --- | ---
                      rsdtdsp | 住居表示 | N番N号
                      parcel | 地番 | N番地
                      undetermined | 判別不能 | N番N（号なし）、N番、N-N、N番地N号
                      unknown | パターンなし | 上記いずれにもマッチしない
                    enum: ["rsdtdsp", "parcel", "undetermined", "unknown"]
                    example: "rsdtdsp"
                required:
                  - input
                  - output
                  - type
              examples:
                type_determined:
                  summary: 住所タイプが判別できる例
                  value:
                    input: "東京都千代田区紀尾井町１番３号東京ガーデンテラス紀尾井町"
                    output: "東京都千代田区紀尾井町1-3 東京ガーデンテラス紀尾井町"
                    type: "rsdtdsp"
                type_undetermined:
                  summary: 住所タイプが判別できない例
                  value:
                    input: "東京都千代田区紀尾井町１−３東京ガーデンテラス紀尾井町"
                    output: "東京都千代田区紀尾井町1-3 東京ガーデンテラス紀尾井町"
                    type: "undetermined"
        "400":
          $ref: "#/components/responses/badRequest"
        "500":
          $ref: "#/components/responses/internalServerError"
  "/match":
    get:
      summary: 住所マッチング
      description: 住所をABRデータと照合し、マッチしたレコードを返します。
      operationId: matchAddress
      parameters:
        - $ref: "#/components/parameters/address"
        - $ref: "#/components/parameters/category"
        - $ref: "#/components/parameters/pref"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          $ref: "#/components/responses/matchOk"
        "400":
          $ref: "#/components/responses/badRequest"
        "500":
          $ref: "#/components/responses/internalServerError"
  "/validate":
    get:
      summary: 住所存在チェック
      description: |
        入力住所をABRデータと照合し、公式の住居番号または地番が存在するかを返します。
        `exists` は住居番号 (`rsdtdsp_rsdt`) または地番 (`parcel`) が見つかった場合に true です。
        `exact` はさらに入力全体がマッチした場合に true です。建物名・部屋番号はABRの対象外なので、
        住所基底が存在しても `status` が `base_address_only` になることがあります。
        HTTP 200は検索処理の成功を示すだけなので、判定には必ず `exists` と `status` を使用してください。
      operationId: validateAddress
      parameters:
        - $ref: "#/components/parameters/address"
        - $ref: "#/components/parameters/category"
        - $ref: "#/components/parameters/pref"
      responses:
        "200":
          description: 住所存在チェック結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateResponse"
        "400":
          $ref: "#/components/responses/badRequest"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/serviceUnavailable"
  "/geocode":
    get:
      summary: ジオコーディング
      description: 住所をジオコーディングします。
      operationId: geocodeAddress
      parameters:
        - $ref: "#/components/parameters/address"
        - $ref: "#/components/parameters/category"
        - $ref: "#/components/parameters/pref"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          $ref: "#/components/responses/geocodeOk"
        "400":
          $ref: "#/components/responses/badRequest"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/serviceUnavailable"
  "/reverse":
    get:
      summary: 逆ジオコーディング
      description: |
        緯度経度から住所情報を取得します。
        ※実験的なエンドポイントです。今後仕様が変わる可能性があります。
      operationId: reverseGeocodeCoordinates
      parameters:
        - $ref: "#/components/parameters/longitude"
        - $ref: "#/components/parameters/latitude"
        - $ref: "#/components/parameters/category"
        - $ref: "#/components/parameters/pref"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          $ref: "#/components/responses/reverseGeocodeOk"
        "400":
          $ref: "#/components/responses/badRequest"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/serviceUnavailable"
...
components:
  schemas:
    MatchLevelEnum:
      type: string
      description: |
        マッチングレベル
        level | description
        --- | ---
        unknown | 不明
        pref | 都道府県
        city | 郡 / 市区町村
        machiaza | 大字・町
        machiaza_detail | 丁目・小字
        rsdtdsp_blk | 街区符号
        rsdtdsp_rsdt | 住居番号
        parcel | 地番
      enum:
        - unknown
        - pref
        - city
        - machiaza
        - machiaza_detail
        - rsdtdsp_blk
        - rsdtdsp_rsdt
        - parcel
    ValidateResponse:
      type: object
      properties:
        exists:
          type: boolean
          description: 公式の住居番号または地番が見つかったか
          example: true
        exact:
          type: boolean
          description: 入力全体が公式住所としてマッチしたか
          example: true
        status:
          type: string
          enum: [exact, base_address_only, insufficient_detail, not_found]
          description: |
            判定結果。`base_address_only` は住所基底は存在するが建物名など未照合部分がある状態、
            `insufficient_detail` は都道府県・市区町村・町字・街区までしか指定されていない状態です。
          example: exact
        input_address:
          type: string
          example: "東京都千代田区紀尾井町1-3"
        matched_address:
          type: string
          example: "東京都千代田区紀尾井町1-3"
        unmatched_address:
          type: array
          nullable: true
          items:
            type: string
          example: null
        match_level:
          $ref: "#/components/schemas/MatchLevelEnum"
        ids:
          $ref: "#/components/schemas/Ids"
        structured_address:
          $ref: "#/components/schemas/StructuredAddress"
        result_info:
          $ref: "#/components/schemas/ResultInfo"
      required:
        - exists
        - exact
        - status
        - input_address
        - matched_address
        - unmatched_address
        - match_level
        - ids
        - structured_address
        - result_info
    Ids:
      type: object
      description: ID情報
      properties:
        lg_code:
          type: string
          nullable: true
          description: 全国地方公共団体コード
          example: "131016"
        machiaza_id:
          type: string
          nullable: true
          description: 町字ID
          example: "0056000"
        rsdt_addr_flg:
          type: string
          nullable: true
          enum: ["1", "0", null]
          description: |
            住居表示フラグ。
            - `1`: 住居表示実施
            - `0`: 住居表示非実施
            - `null`: match_level が machiaza または machiaza_detail で、実施・非実施のいずれか判定できない場合
          example: "1"
        blk_id:
          type: string
          nullable: true
          description: 街区ID
          example: "001"
        rsdt_id:
          type: string
          nullable: true
          description: 住居ID
          example: "003"
        rsdt2_id:
          type: string
          nullable: true
          description: 住居2ID
          example: null
        prc_id:
          type: string
          nullable: true
          description: 地番ID
          example: null
    ResultInfo:
      type: object
      description: 結果に関する情報
      properties:
        count:
          type: integer
          description: 結果の件数
          example: 2
        limit:
          type: integer
          description: 指定された最大件数
          example: 2
        api_version:
          type: string
          description: APIバージョン
          example: "3.0.22"
        db_version:
          type: string
          description: DBバージョン
          example: "3.0.18"
        enabled_category:
          type: string
          description: 有効なカテゴリ
          example: "all"
        enabled_pref:
          type: string
          description: 有効な都道府県（単一の都道府県コードまたは"all"）
          example: "all"
        duration_ms:
          type: number
          format: double
          description: 処理時間（ミリ秒）
          example: 12.5
    StructuredAddress:
      type: object
      description: 構造化住所情報
      properties:
        pref:
          type: string
          nullable: true
          description: 都道府県
          example: "東京都"
        county:
          type: string
          nullable: true
          description: 郡
          example: null
        city:
          type: string
          nullable: true
          description: 市町村
          example: "千代田区"
        ward:
          type: string
          nullable: true
          description: 政令市区
          example: null
        kyoto_st:
          type: string
          nullable: true
          description: 京都通り名
          example: null
        oaza_cho:
          type: string
          nullable: true
          description: "大字・町"
          example: "紀尾井町"
        chome:
          type: string
          nullable: true
          description: 丁目
          example: null
        koaza:
          type: string
          nullable: true
          description: 小字
          example: null
        machiaza_dist:
          type: string
          nullable: true
          description: 同一町字識別情報
          example: null
        blk_num:
          type: string
          nullable: true
          description: 街区符号
          example: "1"
        rsdt_num:
          type: string
          nullable: true
          description: 住居番号
          example: "3"
        rsdt_num2:
          type: string
          nullable: true
          description: 住居番号2
          example: null
        prc_num1:
          type: string
          nullable: true
          description: 地番1
          example: null
        prc_num2:
          type: string
          nullable: true
          description: 地番2
          example: null
        prc_num3:
          type: string
          nullable: true
          description: 地番3
          example: null
  parameters:
    address:
      name: address
      in: query
      description: |
        処理する住所。
        UTF-8のパーセントエンコーディングにも対応。
        最大100文字。
      required: true
      schema:
        type: string
        maxLength: 100
        example: "東京都千代田区紀尾井町1-3 東京ガーデンテラス紀尾井町 19階、20階"
    limit:
      name: limit
      in: query
      description: |
        結果の最大数を指定します。
        デフォルトは1件です。
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
        maximum: 5
        example: 2
    category:
      name: category
      in: query
      example: "all"
      description: |
        検索対象を指定します（省略時はサーバーの enabled_category 設定に従う）。
        category | description
        --- | ---
        all | basic, rsdtdsp, parcel のすべてを対象に検索を行います。
        basic | 都道府県・市区町村・町字を対象に検索を行います。
        rsdtdsp | 住居表示（街区符号・住居番号）を対象に検索を行います。
        parcel | 地番を対象に検索を行います。
      schema:
        type: string
        enum:
          - all
          - basic
          - rsdtdsp
          - parcel
    latitude:
      name: lat
      in: query
      description: "緯度"
      required: true
      schema:
        type: number
        format: double
        minimum: -90.0
        maximum: 90.0
        example: 35.679107172
    longitude:
      name: lon
      in: query
      description: "経度"
      required: true
      schema:
        type: number
        format: double
        minimum: -180.0
        maximum: 180.0
        example: 139.736394597
    pref:
      name: pref
      in: query
      description: |
        検索対象の行政区域を制限します。
        データベース設定の enabled_pref と一致する値のみ指定可能です。
        - 例: `all` → すべての都道府県
        - 例: `13` → 東京都のみ
      required: false
      schema:
        type: string
        example: "all"
  responses:
...
    matchOk:
      description: 住所マッチング結果
      content:
        application/json:
          schema:
            type: object
            properties:
              type:
                type: string
                enum: ["MatchResult"]
              query:
                type: object
                properties:
                  address:
                    type: string
                    description: マッチング対象の住所
                    example: "東京都千代田区紀尾井町1番3号 東京ガーデンテラス紀尾井町 19階、20階"
                  category:
                    type: string
                    description: 検索カテゴリ
                    example: "all"
                  pref:
                    type: string
                    description: 検索対象地域
                    example: "all"
                  limit:
                    type: integer
                    description: 指定された最大結果数
                    example: 2
              result_info:
                $ref: "#/components/schemas/ResultInfo"
              features:
                type: array
                items:
                  type: object
                  properties:
                    matched_address:
                      type: string
                      description: マッチした住所
                      example: "東京都千代田区紀尾井町1-3"
                    unmatched_address:
                      type: array
                      nullable: true
                      items:
                        type: string
                      description: マッチしなかった文字列
                      example: ["東京ガーデンテラス紀尾井町", "19階、20階"]
                    match_level:
                      allOf:
                        - $ref: "#/components/schemas/MatchLevelEnum"
                      example: "rsdtdsp_rsdt"
                    score:
                      type: number
                      format: double
                      minimum: -1
                      maximum: 1
                      description: |
                        候補の並べ替えに使う参考値。match_level が unknown のときは -1。
                        マッチの正しさは表さないため、採否の判断には使わないでください。
                      example: 1
                    ids:
                      $ref: "#/components/schemas/Ids"
                    structured_address:
                      $ref: "#/components/schemas/StructuredAddress"
          examples:
            match_success:
              summary: マッチング結果例
              value:
                type: "MatchResult"
                query:
                  address: "東京都千代田区紀尾井町1番3号 東京ガーデンテラス紀尾井町 19階、20階"
                  category: "rsdtdsp"
                  pref: "all"
                  limit: 2
                result_info:
                  count: 2
                  limit: 2
                  api_version: "3.0.22"
                  db_version: "3.0.18"
                  enabled_category: "all"
                  enabled_pref: "all"
                  duration_ms: 35.69
                features:
                  - matched_address: "東京都千代田区紀尾井町1-3"
                    unmatched_address:
                      - "東京ガーデンテラス紀尾井町"
                      - "19階、20階"
                    match_level: "rsdtdsp_rsdt"
                    score: 1
                    ids:
                      lg_code: "131016"
                      machiaza_id: "0056000"
                      rsdt_addr_flg: "1"
                      blk_id: "001"
                      rsdt_id: "003"
                      rsdt2_id: null
                      prc_id: null
                    structured_address:
                      pref: "東京都"
                      county: null
                      city: "千代田区"
                      ward: null
                      kyoto_st: null
                      oaza_cho: "紀尾井町"
                      chome: null
                      koaza: null
                      machiaza_dist: null
                      blk_num: "1"
                      rsdt_num: "3"
                      rsdt_num2: null
                      prc_num1: null
                      prc_num2: null
                      prc_num3: null
                  - matched_address: "東京都千代田区紀尾井町"
                    unmatched_address: null
                    match_level: "machiaza"
                    score: 1
                    ids:
                      lg_code: "131016"
                      machiaza_id: "0056000"
                      rsdt_addr_flg: "1"
                      blk_id: null
                      rsdt_id: null
                      rsdt2_id: null
                      prc_id: null
                    structured_address:
                      pref: "東京都"
                      county: null
                      city: "千代田区"
                      ward: null
                      kyoto_st: null
                      oaza_cho: "紀尾井町"
                      chome: null
                      koaza: null
                      machiaza_dist: null
                      blk_num: null
                      rsdt_num: null
                      rsdt_num2: null
                      prc_num1: null
                      prc_num2: null
                      prc_num3: null
    badRequest:
      description: "不正なリクエスト"
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                description: エラー状態
                example: "error"
              message:
                type: string
                description: エラーメッセージ
                example: "The address parameter is empty"
            required:
              - status
              - message
    internalServerError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                description: エラー状態
                example: "error"
              message:
                type: string
                description: エラーメッセージ
                example: "Internal Server Error"
            required:
              - status
              - message
    serviceUnavailable:
      description: |
        必要なデータが利用できません。
        - 座標データ無効（enable_pos=false に設定されている）
        - 指定カテゴリのデータがキャッシュに未搭載（例: "parcel data not available in current cache"）
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                description: エラー状態
                example: "error"
              message:
                type: string
                description: エラーメッセージ
                example: "This endpoint requires enable_pos=true in the database."
            required:
              - status
              - message